Move some class name fields into another file
