+++ src/main/kotlin/net/theevilreaper/dartpoet/type/TypeNames.kt
@@ -1,0 +1,19 @@
+package net.theevilreaper.dartpoet.type
+
+@JvmField
+val BOOLEAN: ClassName = ClassName("bool")
+
+@JvmField
+val INTEGER: ClassName = ClassName("int")
+
+@JvmField
+val DOUBLE: ClassName = ClassName("double")
+
+@JvmField
+val STRING: ClassName = ClassName("String")
+
+@JvmField
+val CONST: ClassName = ConstClassName()
+
+@JvmField
+val DYNAMIC: ClassName = DynamicClassName()
